% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{death_distribution}
\alias{death_distribution}
\title{death_distribution
Performs back distribution of data dumps}
\usage{
death_distribution(
  date,
  incidD,
  pois_out,
  use_cases = FALSE,
  incidI = incidD,
  support = FALSE,
  diff = rep(0, length(incidD)),
  out_csse = rep(0, length(incidD)),
  low_date = rep(NA, length(incidD))
)
}
\arguments{
\item{date}{Vector containing dates across which the analysis has to be performed}

\item{incidD}{Vector containing reported deaths on each of the above mentioned dates}

\item{pois_out}{Vector containing indicators if incidD point at date is an outlier 1 or not 0}

\item{use_cases}{TRUE/FALSE to indicate if cases have to be used to back distribute default FALSE}

\item{incidI}{Vector containing reported cases on each of the above mentioned dates if use_cases is TRUE}

\item{support}{TRUE/FALSE to indicate if additional support data for outliers, extra cases and back distribution date is provided}

\item{diff}{Vector containing the number of extra deaths reported on a date in supporting data integer}

\item{low_date}{Vector containing information about the period across which the diff should be back distributed to date type}

\item{out_sup}{Vector indicating if reported incidD at a date is an outlier by supporting data 1 or 0}
}
\value{
Vector of length date indicating the wave number to which data points in incidD belong to
}
\description{
death_distribution
Performs back distribution of data dumps
}
