% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{outlier_detection}
\alias{outlier_detection}
\title{outlier_detection
Detect outlier using Combination 3 method}
\usage{
outlier_detection(date, incidD)
}
\arguments{
\item{date}{Vector containing dates across which the analysis has to be performed}

\item{incidD}{Vector containing reported deaths on each of the above mentioned dates}
}
\value{
Vector of length date indicating which data points in incidD are outliers
}
\description{
outlier_detection
Detect outlier using Combination 3 method
}
